requirements.txt
# Dépendances requises pour ForensicHunter

# Dépendances principales
click>=8.1.3
rich>=13.3.5
tqdm>=4.65.0
psutil>=5.9.5
yara-python>=4.3.0
python-registry>=1.3.1
volatility3>=2.4.1
pywin32>=306; platform_system == "Windows"
wmi>=1.5.1; platform_system == "Windows"
pefile>=2023.2.7
lxml>=4.9.2
jinja2>=3.1.2
pandas>=2.0.1
matplotlib>=3.7.1
networkx>=3.1
python-evtx>=0.7.4
python-magic-bin>=0.4.14; platform_system == "Windows"
python-magic>=0.4.27; platform_system != "Windows"
requests>=2.30.0
aiohttp>=3.8.4
pillow>=9.5.0
colorama>=0.4.6
tabulate>=0.9.0
pyinstaller>=5.10.1; platform_system == "Windows"

# Dépendances pour l'analyse de navigateurs
browsercookie>=0.7.7
pycryptodome>=3.18.0

# Dépendances pour l'analyse mémoire
pymem>=1.12.0
construct>=2.10.68

# Dépendances pour la détection d'anomalies
scikit-learn>=1.2.2
scipy>=1.10.1

# Dépendances pour les tests
pytest>=7.3.1
pytest-cov>=4.1.0
